Fixing hard-coded webpage title and moving content.json higher
