+++ src/content.json
@@ -1,6 +1,6 @@
 {
 	"name": "Mahesh Medam",
-	"about" : "<h2 style='display:inline;'>Mahesh Medam</h2> is an Indian Software Developer looking for the next billion-user project!! He develops software that is convenient to use while preserving aesthetics. An ever-consuming passion for building solutions and exploring new things drives him.<br /><br /> He has been working as a full-time software developer since 2020. Even as a student, he took up roles as an intern/freelancer garnering experience in UI/UX, Backend, Machine learning, and Data. He showed the potential to develop novel solutions by performing in various hackathons. In his free time, you can find him playing games, watching anime, and working on passion projects.<br /> <br /> Check out his work below and if you feel you can have a great time building stuff together, contact him <a href='#contactSection' style='text-decoration:none;'>here👇</a>",
+	"about" : "<b style='display:inline;'>Mahesh Medam</b> is an Indian Software Developer looking for the next billion-user project!! He develops software that is convenient to use while preserving aesthetics. An ever-consuming passion for building solutions and exploring new things drives him.<br /><br /> He has been working as a full-time software developer since 2020. Even as a student, he took up roles as an intern/freelancer garnering experience in UI/UX, Backend, Machine learning, and Data. He showed the potential to develop novel solutions by performing in various hackathons. In his free time, you can find him playing games, watching anime, and working on passion projects.<br /> <br /> Check out his work below and if you feel you can have a great time building stuff together, contact him <a href='#contactSection' style='text-decoration:none;'>here👇</a>",
 	"profileURL":"https://raw.githubusercontent.com/memahesh/myfolio/master/assets/images/profile.png",
 	"resumeURL":"https://drive.google.com/file/d/1chHzWcyUW5wjqwuBPTOHEY6FtnrK0rna/view?usp=sharing",
 	"email":"memahesh219@gmail.com",
